Use p-throttle module for reporting current state This simplifies the logic of throttling reports plus adds the benefit of allowing reports requested during a pending report to still be called after the pending report completes.
Closes: #1945
Change-type: minor
