+++ package.json
@@ -31,6 +31,7 @@
     "@balena/happy-eyeballs": "0.0.6",
     "dbus": "^1.0.7",
     "mdns-resolver": "^1.0.0",
+    "p-throttle": "^4.1.1",
     "semver": "^7.3.2",
     "sinon": "^11.1.2",
     "sqlite3": "^4.1.1",
