+++ src/api-binder/report.ts
@@ -2,24 +2,21 @@ import * as url from 'url';
 import * as _ from 'lodash';
 import { delay } from 'bluebird';
 import { CoreOptions } from 'request';
-import { performance } from 'perf_hooks';
 
 import * as constants from '../lib/constants';
-import { withBackoff, OnFailureInfo } from '../lib/backoff';
 import { log } from '../lib/supervisor-console';
 import { InternalInconsistencyError, StatusError } from '../lib/errors';
 import { getRequestInstance } from '../lib/request';
-
 import { DeviceState } from '../types';
 import * as config from '../config';
 import { SchemaTypeKey, SchemaReturn } from '../config/schema-type';
 import * as eventTracker from '../event-tracker';
 import * as deviceState from '../device-state';
-
 import { shallowDiff, prune, empty } from '../lib/json';
 
+import { OnRetry, AbortLimit, withBackoff, withThrottle } from '../lib/limit';
+
 let lastReport: DeviceState = {};
-let lastReportTime: number = -Infinity;
 export let stateReportErrors = 0;
 
 type StateReportOpts = {
@@ -67,27 +64,16 @@ async function report({ body, opts }: StateReport) {
 async function reportCurrentState(opts: StateReportOpts) {
 	// Wrap the report with fetching of state so report always has the latest state diff
 	const getStateAndReport = async () => {
-		const now = performance.now();
-		// Only try to report if enough time has elapsed since last report
-		if (now - lastReportTime >= constants.maxReportFrequency) {
-			const currentState = await deviceState.getCurrentForReport(lastReport);
-			const stateDiff = prune(shallowDiff(lastReport, currentState, 2));
+		const currentState = await deviceState.getCurrentForReport(lastReport);
+		const stateDiff = prune(shallowDiff(lastReport, currentState, 2));
 
-			if (empty(stateDiff)) {
-				return;
-			}
-
-			await report({ body: stateDiff, opts });
-			lastReportTime = performance.now();
-			lastReport = currentState;
-			log.info('Reported current state to the cloud');
-		} else {
-			// Not enough time has elapsed since last report
-			// Delay report until next allowed time
-			const timeSinceLastReport = now - lastReportTime;
-			await delay(constants.maxReportFrequency - timeSinceLastReport);
-			await getStateAndReport();
+		if (empty(stateDiff)) {
+			throw new AbortLimit();
 		}
+
+		await report({ body: stateDiff, opts });
+		lastReport = currentState;
+		log.info(`${new Date()} - Reported current state to the cloud`);
 	};
 
 	// Create a report that will backoff on errors
@@ -100,13 +86,18 @@ async function reportCurrentState(opts: StateReportOpts) {
 	// Run in try block to avoid throwing any exceptions
 	try {
 		await reportWithBackoff();
-		stateReportErrors = 0;
 	} catch (e) {
-		log.error(e);
+		if (e instanceof AbortLimit) {
+			// Throw AbortLimit because reportCurrentState uses a throttle
+			// this aborts any pending reports so that the next call runs right away
+			throw e;
+		} else {
+			log.error(e);
+		}
 	}
 }
 
-function handleRetry(retryInfo: OnFailureInfo) {
+function handleRetry(retryInfo: OnRetry) {
 	if (retryInfo.error instanceof StatusError) {
 		// We don't want these errors to be classed as a report error, as this will cause
 		// the watchdog to kill the supervisor - and killing the supervisor will
@@ -138,15 +129,18 @@ export async function startReporting() {
 		'appUpdatePollInterval',
 	])) as StateReportOpts;
 
-	let reportPending = false;
+	const throttledReport = withThrottle(
+		reportCurrentState,
+		constants.maxReportFrequency,
+	);
+
 	const doReport = async () => {
-		if (!reportPending) {
-			reportPending = true;
-			await reportCurrentState(reportConfigs);
-			reportPending = false;
+		try {
+			await throttledReport(reportConfigs);
+		} catch (e) {
+			// noop
 		}
 	};
-
 	// If the state changes, report it
 	deviceState.on('change', doReport);
